+++ include/Driver/3DS/PlayerWav3DS.hpp
@@ -17,7 +17,7 @@ namespace SuperHaxagon {
 		void play() override;
 		void pause() override {};
 		bool isDone() const override;
-		double getTime() const override;
+		float getTime() const override;
 
 	private:
 		u8* _data;
