+++ include/Driver/Font.hpp
@@ -19,9 +19,9 @@ namespace SuperHaxagon {
 		Font(Font&) = delete;
 		virtual ~Font() = default;
 
-		virtual void setScale(double scale) = 0;
-		virtual double getHeight() const = 0;
-		virtual double getWidth(const std::string& str) const = 0;
+		virtual void setScale(float scale) = 0;
+		virtual float getHeight() const = 0;
+		virtual float getWidth(const std::string& str) const = 0;
 		virtual void draw(const Color& color, const Point& position, Alignment alignment, const std::string& text) = 0;
 	};
 }
