+++ include/Driver/Platform.hpp
@@ -35,7 +35,7 @@ namespace SuperHaxagon {
 		virtual ~Platform() = default;
 
 		virtual bool loop() = 0;
-		virtual double getDilation() = 0;
+		virtual float getDilation() = 0;
 
 		virtual std::string getPath(const std::string& partial) = 0;
 		virtual std::string getPathRom(const std::string& partial) = 0;
