+++ include/Driver/SFML/FontSFML.hpp
@@ -11,20 +11,20 @@ namespace SuperHaxagon {
 
 	class FontSFML : public Font {
 	public:
-		FontSFML(PlatformSFML& platform, const std::string& path, double size);
+		FontSFML(PlatformSFML& platform, const std::string& path, float size);
 		~FontSFML() override;
 
-		void setScale(double scale) override;
-		double getHeight() const override;
-		double getWidth(const std::string& text) const override;
+		void setScale(float scale) override;
+		float getHeight() const override;
+		float getWidth(const std::string& text) const override;
 		void draw(const Color& color, const Point& position, Alignment alignment, const std::string& text) override;
 
 	private:
 		PlatformSFML& _platform;
 		sf::Font _font;
 		bool _loaded = false;
-		double _scale;
-		double _size;
+		float _scale;
+		float _size;
 	};
 }
