+++ include/Factories/Level.hpp
@@ -18,59 +18,59 @@ namespace SuperHaxagon {
 
 	class Level {
 	public:
-		static constexpr double DIFFICULTY_SCALAR_WALLS = 0.0375;
-		static constexpr double DIFFICULTY_SCALAR_ROT = 0.08;
-		static constexpr double FLIP_FRAMES_MIN = 120;
-		static constexpr double FLIP_FRAMES_MAX = 600;
-		static constexpr double FRAMES_PER_CHANGE_SIDE = 50;
-		static constexpr double FRAMES_PER_SPIN = 90;
-		static constexpr double FRAMES_PER_PULSE = 10;
-		static constexpr double SPIN_SPEED = TAU / 130.0;
-		static constexpr double ROTATE_ZERO_SPEED = TAU / 200.0;
-		static constexpr double PULSE_DISTANCE = 5.0;
+		static constexpr float DIFFICULTY_SCALAR_WALLS = 0.0375f;
+		static constexpr float DIFFICULTY_SCALAR_ROT = 0.08f;
+		static constexpr float FLIP_FRAMES_MIN = 120;
+		static constexpr float FLIP_FRAMES_MAX = 600;
+		static constexpr float FRAMES_PER_CHANGE_SIDE = 50;
+		static constexpr float FRAMES_PER_SPIN = 90;
+		static constexpr float FRAMES_PER_PULSE = 10;
+		static constexpr float SPIN_SPEED = TAU / 130.0f;
+		static constexpr float ROTATE_ZERO_SPEED = TAU / 200.0f;
+		static constexpr float PULSE_DISTANCE = 5.0f;
 		static constexpr int MIN_SAME_SIDES = 3;
 		static constexpr int MAX_SAME_SIDES = 5;
 
-		Level(const LevelFactory& factory, Twist& rng, double patternDistCreate);
+		Level(const LevelFactory& factory, Twist& rng, float patternDistCreate);
 		Level(Level&) = delete;
 		~Level();
 
-		void update(Twist& rng, double patternDistDelete, double patternDistCreate, double dilation);
-		void draw(Game& game, double scale, double offsetWall) const;
-		Movement collision(double cursorDistance, double dilation) const;
+		void update(Twist& rng, float patternDistDelete, float patternDistCreate, float dilation);
+		void draw(Game& game, float scale, float offsetWall) const;
+		Movement collision(float cursorDistance, float dilation) const;
 
 		void increaseMultiplier();
 		void clearPatterns();
-		void rotate(double distance, double dilation);
-		void left(double dilation);
-		void right(double dilation);
+		void rotate(float distance, float dilation);
+		void left(float dilation);
+		void right(float dilation);
 		void clamp();
 
 		// Effects
 		void spin();
 		void invertBG();
-		void pulse(double scale);
+		void pulse(float scale);
 
 		// Time
-		double getFrame() const {return _frame;}
+		float getFrame() const {return _frame;}
 
 		const LevelFactory& getLevelFactory() const {return *_factory;}
 
 		// Stuff for Win control
 		std::deque<Pattern>& getPatterns() {return _patterns;}
-		void setWinMultiplierRot(const double multiplier) {_multiplierRot = multiplier;}
-		void setWinMultiplierWalls(const double multiplier) {_multiplierWalls = multiplier;}
+		void setWinMultiplierRot(const float multiplier) {_multiplierRot = multiplier;}
+		void setWinMultiplierWalls(const float multiplier) {_multiplierWalls = multiplier;}
 		void setWinAutoPatternCreate(const bool autoPatternCreate) {_autoPatternCreate = autoPatternCreate;}
 		void setWinShowCursor(const bool show) {_showCursor = show;}
-		void setWinFrame(const double frame) {_frame = frame;}
+		void setWinFrame(const float frame) {_frame = frame;}
 		void setWinRotationToZero() {_rotateToZero = true;}
 		void setWinFactory(const LevelFactory* factory);
 		void setWinSides(int sides);
 		void resetColors();
 
 	private:
-		void advanceWalls(Twist& rng, double patternDistDelete, double patternDistCreate);
-		void reverseWalls(Twist& rng, double patternDistDelete, double patternDistCreate);
+		void advanceWalls(Twist& rng, float patternDistDelete, float patternDistCreate);
+		void reverseWalls(Twist& rng, float patternDistDelete, float patternDistCreate);
 		const PatternFactory& getRandomPattern(Twist& rng);
 		
 		const LevelFactory* _factory;
@@ -81,22 +81,22 @@ namespace SuperHaxagon {
 		bool _showCursor = true;
 		bool _rotateToZero = false;
 		
-		double _multiplierRot = 0.9; // Current direction and speed of rotation
-		double _multiplierWalls = 0.85; // Current speed of the walls flying at you
-		double _cursorPos{};
-		double _rotation{};
-		double _sidesTween{};
+		float _multiplierRot = 0.9f; // Current direction and speed of rotation
+		float _multiplierWalls = 0.85f; // Current speed of the walls flying at you
+		float _cursorPos{};
+		float _rotation{};
+		float _sidesTween{};
 
 		// Side change mechanics
 		int _sidesLast{}; // The sides that we are transitioning FROM
 		int _sidesCurrent{}; // The sides we are transitioning TO
 
 		// Event timings
-		double _frame{}; // Frame this level is on
-		double _delayFrame{}; // Tween between side switches
-		double _delayMax{};   // When a delay starts, this is the initial value of _delayFrame
-		double _tweenFrame{}; // Tween colors
-		double _flipFrame = FLIP_FRAMES_MAX; // Amount of frames left until it rotates in the opposite direction
+		float _frame{}; // Frame this level is on
+		float _delayFrame{}; // Tween between side switches
+		float _delayMax{};   // When a delay starts, this is the initial value of _delayFrame
+		float _tweenFrame{}; // Tween colors
+		float _flipFrame = FLIP_FRAMES_MAX; // Amount of frames left until it rotates in the opposite direction
 
 		std::map<LocColor, Color> _color;
 		std::map<LocColor, Color> _colorNext;
@@ -105,9 +105,9 @@ namespace SuperHaxagon {
 		int _sameCount = 0; // When 0, allows the level to select any pattern instead of currentSides
 		int _sameSides = 0; // Sides of the last selected pattern
 		bool _bgInverted = false;
-		double _pulse = 0.0;
-		double _spin = 0.0;
-		double _frontGap = 0.0;
+		float _pulse = 0.0;
+		float _spin = 0.0;
+		float _frontGap = 0.0;
 	};
 
 	class LevelFactory {
@@ -118,7 +118,7 @@ namespace SuperHaxagon {
 		LevelFactory(std::ifstream& file, std::vector<std::shared_ptr<PatternFactory>>& shared, LocLevel location, Platform& platform, size_t levelIndexOffset);
 		LevelFactory(const LevelFactory&) = delete;
 
-		std::unique_ptr<Level> instantiate(Twist& rng, double renderDistance) const;
+		std::unique_ptr<Level> instantiate(Twist& rng, float renderDistance) const;
 
 		bool isLoaded() const {return _loaded;}
