+++ include/States/Over.hpp
@@ -13,18 +13,18 @@ namespace SuperHaxagon {
 
 	class Over : public State {
 	public:
-		static constexpr double GAME_OVER_ROT_SPEED = TAU/400.0;
-		static constexpr double GAME_OVER_ACCELERATION_RATE = 0.1;
+		static constexpr float GAME_OVER_ROT_SPEED = TAU/400.0f;
+		static constexpr float GAME_OVER_ACCELERATION_RATE = 0.1f;
 		static constexpr int FRAMES_PER_GAME_OVER = 60;
 		static constexpr int PULSE_TIME = 75;
 
-		Over(Game& game, std::unique_ptr<Level> level, LevelFactory& selected, double score, std::string text);
+		Over(Game& game, std::unique_ptr<Level> level, LevelFactory& selected, float score, std::string text);
 		Over(Over&) = delete;
 		~Over() override;
 
-		std::unique_ptr<State> update(double dilation) override;
-		void drawTop(double scale) override;
-		void drawBot(double scale) override;
+		std::unique_ptr<State> update(float dilation) override;
+		void drawTop(float scale) override;
+		void drawBot(float scale) override;
 		void enter() override;
 
 	private:
@@ -35,9 +35,9 @@ namespace SuperHaxagon {
 		std::string _text = "GAME OVER";
 
 		bool _high = false;
-		double _score = 0;
-		double _frames = 0;
-		double _offset = 1.0;
+		float _score = 0;
+		float _frames = 0;
+		float _offset = 1.0;
 	};
 }
