+++ include/States/Transition.hpp
@@ -13,16 +13,16 @@ namespace SuperHaxagon {
 
 	class Transition : public State {
 	public:
-		static constexpr double TRANSITION_ACCELERATION_RATE = 0.1;
+		static constexpr float TRANSITION_ACCELERATION_RATE = 0.1f;
 		static constexpr int TRANSITION_FRAMES = 120;
 		
-		Transition(Game& game, std::unique_ptr<Level> level, LevelFactory& selected, double score);
+		Transition(Game& game, std::unique_ptr<Level> level, LevelFactory& selected, float score);
 		Transition(Transition&) = delete;
 		~Transition() override;
 
-		std::unique_ptr<State> update(double dilation) override;
-		void drawTop(double scale) override;
-		void drawBot(double scale) override;
+		std::unique_ptr<State> update(float dilation) override;
+		void drawTop(float scale) override;
+		void drawBot(float scale) override;
 		void enter() override;
 
 	private:
@@ -31,9 +31,9 @@ namespace SuperHaxagon {
 		LevelFactory& _selected;
 		std::unique_ptr<Level> _level;
 
-		double _score = 0;
-		double _frames = 0;
-		double _offset = 1.0;
+		float _score = 0;
+		float _frames = 0;
+		float _offset = 1.0;
 		int _levelIndex = 0;
 	};
 }
