+++ source/Driver/3DS/PlayerWav3DS.cpp
@@ -36,7 +36,7 @@ namespace SuperHaxagon {
 		return _buffer.status == NDSP_WBUF_DONE;
 	}
 
-	double PlayerWav3DS::getTime() const {
+	float PlayerWav3DS::getTime() const {
 		// Don't care about the time of WAV files
 		return 0;
 	}
